Split workspace surface helpers
