+++ docs/DECISIONS.md
@@ -89,7 +89,7 @@
 - Assistant response feedback is recorded as `messageFeedback` thread events keyed by message ID rather than as view-local state. The surface reduces the latest feedback event per message so SwiftUI, static HTML, persisted threads, and Playwright screenshots stay aligned while keeping feedback controls assistant-only.
 - Assistant Retry is shown only on the latest assistant response and dispatches the existing `retry-last-turn` command. This keeps the visible action Codex-like while preserving the normal composer path for slash handling, tool dispatch, safety review, persistence, and cancellation.
 - Workspace check automations share the same deterministic quick-schedule and natural-language schedule parser as thread follow-ups. Visible buttons, command palette entries, and `/workspace-check ...` all persist concrete `nextRunAt` values through one workspace-schedule creation path. Simple recurrence (`hourly`, `daily`, `weekly`, and `every N minutes/hours/days/weeks`) is persisted on the automation record and advances `nextRunAt` after each run.
-- TrustedRouter runtime failures normalize into a shared `RuntimeIssueSurface` rather than renderer-specific strings. The top bar, menu bar, settings sheet, static HTML renderer, and Playwright harness consume the same severity/title/message/action data for missing sign-in, missing developer key, rejected key, rate limits, network failure, empty response, malformed model actions, and generic run failures.
+- TrustedRouter runtime failures normalize into a shared `RuntimeIssueSurface` rather than renderer-specific strings. `WorkspaceRuntimeIssueBuilder` owns failure classification and diagnostics. The top bar, menu bar, settings sheet, static HTML renderer, and Playwright harness consume the same severity/title/message/action data for missing sign-in, missing developer key, rejected key, rate limits, network failure, empty response, malformed model actions, and generic run failures.
 - Runtime issue retry is a command, not a view-only callback. The Retry action prepares the latest non-empty user turn from the selected transcript and submits it through the normal composer path, so slash commands, tool dispatch, safety review, persistence, and future telemetry stay identical to a user pressing Send again.
 - Runtime issue recovery actions should take the user directly to the corrective control. Malformed model/tool responses and provider rate limits use `Switch model` to open the same searchable top-bar model picker and focus model search, instead of introducing a second recovery modal or leaving the user to find the picker manually.
 - Runtime issue diagnostics are generated from current runtime state, not hardcoded per error. Settings shows the API base URL, authentication mode, key state, model, agent status, rate-limit metadata when present, and a redacted last-error snippet when present. Secret-like `sk-...` and bearer tokens are redacted before they reach the surface.
